+++ src/uttaksplan/uttaksplaner/oppsett/forenklet.ts
